Move workshop_mod_updater_panel.py into RimSort.windows
